update code
Development of data preprocessing function for the model that reduces the over-representation of certain steering angles of the vehicles.

This was done by reducing the number of values seen above the average frequency as seen in the generated histograms. Values were dropped out based on a probability using np.random.choice and 2 variables with p=[0.4, 0.6] respectively. If he variable of 60% was chosen the values were dropped out. This avoided dropping the save values as it progressed through a for loop.