final commit
Tuning hyperparameters until the results of the final model and weights satistfied me with the performance in the simulation.

Biggest take away and tuning keypoint was the focus on the 'steering_anglesteering angle correction that is applied to the images as viewed from the side cameras and getting them to lineup more accurately with the center camera. The correction value that produced desired results was 'correction = 0.315'

Additionally some trimming and preprocessing to the initial data was done to further inhance the results.

The source of data that represented me driving the track in reverse in attempt to make the model more diverse was unnecessary and proved to be too much data for the purpose of the simulation.